Describe workaround for long checkout times
When I tried this in https://github.com/dscho/git-sdk-64, the `checkout`
step took a whopping five minutes. We don't need the files, and we don't
even need to transfer all those Git objects: partial clone is here to
save the day.

It looks a bit more complicated than the originally-proposed solution,
so I describe this only as an alternative instead of replacing the
short-and-sweet solution using `actions/checkout@v2`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>